requirements.txt
requests
google-cloud-bigquery
cached_property
pandas
pandas-gbq
google-cloud-storage